// netrika/src/common/interfaces/IAppSettingsRequest.ts
export interface IAppSettingsRequest {
  REACT_APP_ROOT_FOLDER: string;
  REACT_APP_API_URL: string;
  REACT_APP_API_ROUTER_URL: string;
  authority: string;
  authorityImk: string;
  clientId: string;
  clientScope: string;
  responseType: string;
  clientSecret: string;
  ExlabUI_URL: string;
  REACT_APP_SQL_LAB: string;
  show_screen_route: boolean;
  redirect_uri: string;
  redirect_uri_iemk: string;
  region_name: string;
}